Fix incorrect stack calculation
The stack calculation is incorrect. The stack due to fork offset should be moved outside of the group being multiplied by `sin(hta)` because it already accounts for hta.

That being said, stack/reach calculation with this app is already almost guaranteed to be off anyways because it doesn't seem to account for headset stack height? That should probably be added too.